Move fakes to their own folder, refactor EventContextConsumerTest
