+++ src/test/java/com/teragrep/aer_01/fakes/CheckpointlessEventContextFactory.java
@@ -44,7 +44,7 @@
  * a licensee so wish it.
  */
 
-package com.teragrep.aer_01;
+package com.teragrep.aer_01.fakes;
 
 import com.azure.messaging.eventhubs.CheckpointStore;
 import com.azure.messaging.eventhubs.EventData;
