+++ src/test/java/com/teragrep/aer_01/fakes/EventContextFactory.java
@@ -44,35 +44,10 @@
  * a licensee so wish it.
  */
 
-package com.teragrep.aer_01;
+package com.teragrep.aer_01.fakes;
 
-import com.azure.messaging.eventhubs.CheckpointStore;
-import com.azure.messaging.eventhubs.models.Checkpoint;
-import com.azure.messaging.eventhubs.models.PartitionOwnership;
-import reactor.core.publisher.Flux;
-import reactor.core.publisher.Mono;
+import com.azure.messaging.eventhubs.models.EventContext;
 
-import java.util.List;
-
-public class CheckpointStoreFake implements CheckpointStore {
-
-    @Override
-    public Flux<PartitionOwnership> listOwnership(String s, String s1, String s2) {
-        throw new UnsupportedOperationException("CheckpointStoreFake does not implement this function.");
-    }
-
-    @Override
-    public Flux<PartitionOwnership> claimOwnership(List<PartitionOwnership> list) {
-        throw new UnsupportedOperationException("CheckpointStoreFake does not implement this function.");
-    }
-
-    @Override
-    public Flux<Checkpoint> listCheckpoints(String s, String s1, String s2) {
-        throw new UnsupportedOperationException("CheckpointStoreFake does not implement this function.");
-    }
-
-    @Override
-    public Mono<Void> updateCheckpoint(Checkpoint checkpoint) {
-        return Mono.empty();
-    }
+public interface EventContextFactory {
+    EventContext create();
 }
